requirements.txt
matplotlib==3.6.2
Pillow==9.5.0
pyowm==3.2.0
qrcode==7.3.1
spidev==3.5
